+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # CHANGELOG
 
 
+**v0.2.0 - 04-29-2026**
+---
+
+- Refactored N+1 query on folder switch - [#117](https://github.com/hkdb/aerion/issues/117)
+- Fixed IMAP server side search regression
+- Fixed settings tab icons - [#125](https://github.com/hkdb/aerion/issues/125)
+- Fixed plain text links - [#113](https://github.com/hkdb/aerion/issues/113)
+
+
 **v0.1.39 - 04-27-2026**
 ---